django_main/pyproject.toml
[project]
name = "django-main"
version = "0.1.0"
description = ""
authors = [
    {name = "َAmirhosein",email = "amirhoosenbabai82@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "grpcio (>=1.75.0,<2.0.0)",
    "grpcio-tools (>=1.75.0,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
